mcp_ocr/requirements.txt
# Opcional: OCR real (requer Tesseract OCR no sistema: apt install tesseract-ocr tesseract-ocr-por)
Pillow>=10.0.0
pytesseract>=0.3.10
